# book-network-ui/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenApi specification",
        "description": "OpenApi Documentation for Spring security",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Clem",
            "url": "http://clem.nz",
            "email": "contact@clem.nz"
        },
        "license": {
            "name": "License name",
            "url": "http://clem.nz/info"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        },
        {
            "url": "clem.nz/api",
            "description": "PROD ENV"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/authenticate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "authenticate",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/activate-account": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "operationId": "confirm",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "RegistrationRequest": {
                "required": [
                    "email",
                    "firstname",
                    "lastname",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 2147483647,
                        "minLength": 8,
                        "type": "string"
                    }
                }
            },
            "AuthenticationRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "maxLength": 2147483647,
                        "minLength": 8,
                        "type": "string"
                    }
                }
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT auth description",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}